YOLOServer/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
Flask>=1.1.2
ultralytics